+++ apps/reminder/src/main/java/it/ioapp/com/reminder/scheduler/CheckRemindersToDeleteJob.java
@@ -1,0 +1,34 @@
+package it.ioapp.com.reminder.scheduler;
+
+import it.ioapp.com.reminder.service.ReminderService;
+import java.time.Duration;
+import java.time.Instant;
+import javax.transaction.Transactional;
+import lombok.extern.slf4j.Slf4j;
+import org.quartz.Job;
+import org.quartz.JobExecutionContext;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+@Component
+@Slf4j
+public class CheckRemindersToDeleteJob implements Job {
+
+  private static final String JOB_LOG_NAME = "Reminders to DELETE Job ";
+
+  private final ReminderService reminderService;
+
+  @Autowired
+  public CheckRemindersToDeleteJob(ReminderService reminderService) {
+    this.reminderService = reminderService;
+  }
+
+  @Transactional(Transactional.TxType.NOT_SUPPORTED)
+  public void execute(JobExecutionContext context) {
+    log.info(JOB_LOG_NAME + "started");
+    Instant start = Instant.now();
+    reminderService.deleteMessage();
+    Instant end = Instant.now();
+    log.info(JOB_LOG_NAME + "ended in " + Duration.between(start, end).getSeconds() + " seconds");
+  }
+}