+++ apps/sending-func/src/functions/Notify/handler.ts
@@ -1,4 +1,4 @@
-import { UserSessionInfo } from "@pagopa/io-backend-session-sdk/UserSessionInfo";
+import { UserSessionInfo } from "@/generated/session-manager/UserSessionInfo";
 import { PushNotificationsContentTypeEnum } from "@pagopa/io-functions-commons/dist/generated/definitions/PushNotificationsContentType";
 import { ReminderStatusEnum } from "@pagopa/io-functions-commons/dist/generated/definitions/ReminderStatus";
 import { RetrievedProfile } from "@pagopa/io-functions-commons/dist/src/models/profile";