+++ infra/resources/_modules/container_apps/variables.tf
@@ -57,7 +57,8 @@ variable "entra_id_admin_ids" {
   description = "Id of Entra ID groups that should be admins of the Container App Environment"
 }
 
-variable "acr_id" {
-  type        = string
-  description = "The Id of the ACR to pull images from"
+variable "application_insights" {
+  type = object({
+    connection_string = string
+  })
 }